requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
django-etc==1.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-hitcount==1.3.3
django-taggit==1.5.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
